+++ playground/src/App.vue
@@ -159,7 +159,7 @@ const actionsList = [
   {
     label: "启动服务",
     icon: "status-running",
-    class: "app-btn-start",
+    className: "app-btn-start",
     handler: () => {
       addToast("success", "指令已发送", "服务启动指令下发成功。");
     },
@@ -167,7 +167,7 @@ const actionsList = [
   {
     label: "暂停服务",
     icon: "settings",
-    class: "app-btn-stop",
+    className: "app-btn-stop",
     handler: () => {
       addToast("warning", "指令已发送", "服务暂停指令下发成功。");
     },
@@ -175,7 +175,7 @@ const actionsList = [
   {
     label: "删除容器",
     icon: "trash",
-    class: "app-btn-delete",
+    className: "app-btn-delete",
     handler: () => {
       addToast("error", "删除警告", "此操作无法撤销。");
     },